requirements.txt
flask
psycopg[binary]
requests
python-dotenv
psycopg[binary]
python-dotenv